rename files
Rename files to match normal python naming convention
