+++ compute_graphs/graph_5d.py
@@ -1,7 +1,7 @@
 import heterocl as hcl
 import numpy as np
 import time
-from computeGraphs.CustomGraphFunctions import *
+from compute_graphs.custom_graph_functions import *
 from user_definer import *
 
 ##############################  5D DERIVATIVE FUNCTIONS #############################
